Fix linting failure on deploy
Latest version of `isort` does not behave the same depending on which
dependencies are currently installed. On deploy, `top_model` dependencies
were not deployed and so they were not found by `isort`. In this situation,
`isort` force to add an empty line before their `import`s.

This problem did not happen on dev workstations as `top_model` is
usually installed.